MIDI Added
Added CasterScript.py that mutes caster channels using an AKAI MIDI pad. The next stage is to integrate this script into the main server and if possible simultaneously build the server in such a way that all changing variables (ip adressess and channel numbers) are read from an external file so that the same code can be used on different systems.
Also added MIDIDEvices.py which is a script that was used to try out setting up the MIDI pad as an object. It is not being used in the mains scripts at this point, but was added because this might be a good way to generalize everything further.